moved menu functions to module
